`CaptchaClientBehavior` extracted
