+++ framework/captcha/assets/yii.captcha.js
